+++ src/Juniper.Root/Units/Angles/Degrees.cs
@@ -8,29 +8,29 @@ namespace Juniper.Units
         /// <summary>
         /// Conversion factor from circles to degrees.
         /// </summary>
-        public const float PER_CIRCLE = 360f;
+        public const double PER_CIRCLE = 360;
 
         /// <summary>
         /// Conversion factor from radians to degrees.
         /// </summary>
-        public const float PER_RADIAN = PER_CIRCLE / Units.Radians.PER_CIRCLE;
+        public const double PER_RADIAN = PER_CIRCLE / Units.Radians.PER_CIRCLE;
 
         /// <summary>
         /// Conversion factor from gradians to degrees.
         /// </summary>
-        public const float PER_GRADIAN = PER_CIRCLE / Units.Gradians.PER_CIRCLE;
+        public const double PER_GRADIAN = PER_CIRCLE / Units.Gradians.PER_CIRCLE;
 
         /// <summary>
         /// Conversion factor from hours to degrees.
         /// </summary>
-        public const float PER_HOUR = PER_CIRCLE / Units.Hours.PER_DAY;
+        public const double PER_HOUR = PER_CIRCLE / Units.Hours.PER_DAY;
 
         /// <summary>
         /// Convert from degrees to radians.
         /// </summary>
         /// <param name="degrees">The number of degrees</param>
         /// <returns>The number of radians</returns>
-        public static float Radians(float degrees)
+        public static double Radians(double degrees)
         {
             return degrees * Units.Radians.PER_DEGREE;
         }
@@ -40,7 +40,7 @@ namespace Juniper.Units
         /// </summary>
         /// <param name="degrees">The number of degrees</param>
         /// <returns>The number of gradians</returns>
-        public static float Gradians(float degrees)
+        public static double Gradians(double degrees)
         {
             return degrees * Units.Gradians.PER_DEGREE;
         }
@@ -50,7 +50,7 @@ namespace Juniper.Units
         /// </summary>
         /// <param name="degrees">The number of degrees</param>
         /// <returns>The number of hours</returns>
-        public static float Hours(float degrees)
+        public static double Hours(double degrees)
         {
             return degrees * Units.Hours.PER_DEGREE;
         }
